feat:  实现多商户数据隔离,支持使用后台配置控制日志路由

+++ src/Http/routes.php
@@ -2,6 +2,9 @@
 
 use Weiaibaicai\OperationLog\Http\Controllers;
 use Illuminate\Support\Facades\Route;
+use Weiaibaicai\OperationLog\Models\OperationLog;
 
-Route::get('auth/operation-logs', Controllers\LogController::class.'@index')->name('dcat-admin.operation-log.index');
-Route::delete('auth/operation-logs/{id}', Controllers\LogController::class.'@destroy')->name('dcat-admin.operation-log.destroy');
+if (OperationLog::withRoutes()) {
+    Route::get('auth/operation-logs', Controllers\LogController::class.'@index')->name('dcat-admin.operation-log.index');
+    Route::delete('auth/operation-logs/{id}', Controllers\LogController::class.'@destroy')->name('dcat-admin.operation-log.destroy');
+}
